services/clinical-notes/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.47
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
aio-pika==9.5.4
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.2
aiosqlite==0.20.0
